Fix traceback in selective reporting view
This commit fixes an error that arises when rendering the listing
for selective reporting of sensitivity testing results. The
folder item comes by default with a boolean value assigned for the
key "disabled", with an action scope for the whole row. However, we
need specific checkboxes to be disabled, not the whole row.